chore: abstract header with data to its own component
