+++ website/src/app/[lang]/layout.tsx
@@ -12,10 +12,7 @@ import '@styles/main.scss';
 import { ApolloWrapper } from '@app/apollo-wrapper';
 import Footer from '@components/footer';
 import Header from '@components/header';
-import { PreloadQuery } from '@graphql/client';
-import { GetHeaderDocument } from '@graphql/queries/header/index.generated';
 import { AVAILABLE_LOCALES, LocaleParamsPath } from '@i18n/locales';
-import { Suspense } from 'react';
 
 export const metadata = {
   title: 'Welcome to website',
@@ -38,12 +35,7 @@ const RootLayout: React.FC<RootLayoutProps> = ({
     <html lang={lang} className={nunitoSans.variable}>
       <body>
         <ApolloWrapper>
-          <PreloadQuery query={GetHeaderDocument}>
-            {/* biome-ignore lint/complexity/noUselessFragments: <explanation> */}
-            <Suspense fallback={<>loading</>}>
-              <Header />
-            </Suspense>
-          </PreloadQuery>
+          <Header lang={lang} />
           <main>{children}</main>
           <Footer />
         </ApolloWrapper>
